Rename AasxGenerate to AasToolkit
Originally, the program AasxGenerate was created to demonstrate how AASX
can be automatically generated in the C# code. This program grew over
the time and acquired many more features (such as validation).

The time has come when the original name (AasxGenerate) does not do
service to its importance and purpose. Hence we rename it in this change
to a more general name, AasxToolkit, leaning on similar tools in the
Linux universe (*e.g.*, pdftk).
