+++ src/AasxPackageExplorer.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 16.0.30413.136
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AasxPackageExplorer", "AasxPackageExplorer\AasxPackageExplorer.csproj", "{569B369E-9393-4F57-994E-6B84398FD7CC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AasxGenerate", "AasxGenerate\AasxGenerate.csproj", "{294FC59A-5645-412F-8216-702FB66528C1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AasxToolkit", "AasxToolkit\AasxToolkit.csproj", "{294FC59A-5645-412F-8216-702FB66528C1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AasxCsharpLibrary", "AasxCsharpLibrary\AasxCsharpLibrary.csproj", "{9863799B-4E44-4DA2-9120-C85C7985BC6D}"
 EndProject
