+++ src/AasxPackageExplorer/AasxPackageExplorer.csproj
@@ -273,9 +273,9 @@
       <Project>{dec389bc-59bc-48e5-b163-6e44ce782c4a}</Project>
       <Name>AasxDictionaryImport</Name>
     </ProjectReference>
-    <ProjectReference Include="..\AasxGenerate\AasxGenerate.csproj">
+    <ProjectReference Include="..\AasxToolkit\AasxToolkit.csproj">
       <Project>{294fc59a-5645-412f-8216-702fb66528c1}</Project>
-      <Name>AasxGenerate</Name>
+      <Name>AasxToolkit</Name>
     </ProjectReference>
     <ProjectReference Include="..\AasxIntegrationBaseWpf\AasxIntegrationBaseWpf.csproj">
       <Project>{e3ab36ea-e98a-4cc2-bc18-1d0e40eeae1a}</Project>
