+++ src/AasxPackageExplorer/AasxPackageExplorer.options.json
@@ -1,5 +1,5 @@
 {
-  /* "AasxToLoad": "..\\..\\..\\..\\AasxGenerate\\bin\\Debug\\sample.aasx", */
+  /* "AasxToLoad": "..\\..\\..\\..\\AasxToolkit\\bin\\Debug\\sample.aasx", */
   /* "AasxToLoad": "sample-hsu.aasx", */
   /* "AasxToLoad": "MTPsample.aasx", */
   "WindowLeft": -1,
