+++ src/AasxPackageExplorer/options-debug.json
@@ -1,6 +1,6 @@
 {
   /* "AasxToLoad" :  "sample-techdata.aasx", */
-  /* "AasxToLoad": "..\\..\\..\\..\\AasxGenerate\\bin\\Debug\\sample.aasx", */
+  /* "AasxToLoad": "..\\..\\..\\..\\AasxToolkit\\bin\\Debug\\sample.aasx", */
   /* "AasxToLoad": "sample-hsu.aasx", */
   /* "AasxToLoad": "MTPsample.aasx", */
   /* "AasxToLoad": "sample-generated-modified.aasx", */
