+++ src/AasxToolkit/data/README.md
@@ -1,5 +1,5 @@
 ﻿# Binary files
 
 This file is intended to create an (empty) /data folder in the path of the binary executable.
-Within this folder, various necessary files for the successfull operation of AasxGenerate.exe are required.
+Within this folder, various necessary files for the successfull operation of AasxToolkit.exe are required.
 They could be copied manually or by engaging the CopyExternalData.bat batch.
