+++ src/AasxToolkit/schemaV10/SimpleDrehzahl.xml
