+++ src/AasxToolkit/schemaV10/validate_samle_json.bat
