+++ src/AasxToolkit/schemaV10/validate_samle_xml.bat
