+++ src/AasxToolkit/schemaV20/Example_AAS_ServoDCMotor - Simplified V2.0.xml
