+++ src/AasxToolkit/schemaV20/SimpleDrehzahl.xml
