+++ src/AasxToolkit/schemaV20/minimum.xml
