+++ src/AasxToolkit/schemaV20/validate_samle_xml.bat
