+++ src/AasxToolkit/schemaV201/Example_AAS_ServoDCMotor - Simplified V2.0.xml
