+++ src/AasxToolkit/schemaV201/minimum.xml
